Use sudo by default when needed
Prior to this, a user would need to add `--sudo` to use sudo when
installing Puppet agents -- even if they specified a sudo password or a
non-root user. This was cumbersome and added an extra step.

This commit adds some logic to automatically enable the usage of sudo if
the `sudo_password` key is found in the credentials hash or if the
`username` key is something other than the string "root".