Update README with more information, typo fixes
This commit adds a little more detail on the custom script, mentioning that it
must be in the same directory as install.bash, on the master.  This commit also
updates the available CLI flags, to add --debug and --trace, as well as to
remove the '=' after some of the flags, since they want a space instead.