Fix the check for pre-existing install.ps1
Prior to this, there was an extra `$` symbol in the if statement that
checks for an existing install.ps1. While the code still compiled and
worked, it's not proper style to have the extra `$`.

Also added some spec tests to double-check that that if statement is
working and hopefully catch issues and regessions like this in the
future.